+++ internal_filesystem/lib/mpos/__init__.py
@@ -19,6 +19,7 @@ from .device_info import DeviceInfo
 from .build_info import BuildInfo
 from .lora_manager import LoRaManager
 from .ir_manager import IRManager
+from .gps_manager import GPSManager
 
 # Battery manager (imported early for UI dependencies)
 from .battery_manager import BatteryManager
@@ -73,7 +74,7 @@ __all__ = [
     "SharedPreferences",
     "ConnectivityManager", "DownloadManager", "WifiService", "AudioManager", "Intent",
     "ActivityNavigator", "AppManager", "TaskManager", "CameraManager", "BatteryManager", "WebServer",
-    "LoRaManager", "IRManager",
+    "LoRaManager", "IRManager", "GPSManager",
     # Device and build info
     "DeviceInfo", "BuildInfo",
     # Common activities